requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
WTForms==3.1.2

# Production WSGI server
gunicorn==23.0.0  

# Async workers for Flask-SocketIO
eventlet==0.36.1  
gevent==24.2.1
gevent-websocket==0.10.1
